refactor(rpc): split terminal methods into cohesive modules
