+++ src/main/runtime/rpc/methods/terminal/terminal-subscription-stream-id.ts
@@ -1,0 +1,5 @@
+let nextTerminalStreamId = 1
+
+export function allocateTerminalSubscriptionStreamId(): number {
+  return nextTerminalStreamId++
+}
