+++ src/main/runtime/rpc/methods/terminal/terminal-viewport-update.ts
@@ -1,0 +1,38 @@
+import type { OrcaRuntimeService } from '../../../orca-runtime'
+import type { TerminalViewportClient } from './terminal-stream-types'
+
+export async function updateViewportForClient(
+  runtime: OrcaRuntimeService,
+  ptyId: string,
+  subscriptionKey: string,
+  client: TerminalViewportClient,
+  viewport: { cols: number; rows: number },
+  defaultType: 'mobile' | 'desktop',
+  // Why: the one-shot RPC has no disconnect hook, so 'refresh' only updates a stream-owned floor; stream paths that own cleanup 'register'.
+  registration: 'register' | 'refresh' = 'register',
+  claim = false
+): Promise<{ updated: boolean; applied: boolean }> {
+  const type = client.type ?? defaultType
+  if (type === 'mobile') {
+    return runtime.updateMobileViewport(ptyId, client.id, viewport)
+  }
+  // Why: stream attachment observes geometry without taking control; a later claim frame makes it authoritative.
+  const updated =
+    registration === 'refresh'
+      ? await runtime.refreshRemoteDesktopViewer(
+          ptyId,
+          client.id,
+          viewport.cols,
+          viewport.rows,
+          claim
+        )
+      : await runtime.updateRemoteDesktopViewer(
+          ptyId,
+          subscriptionKey,
+          client.id,
+          viewport.cols,
+          viewport.rows,
+          claim
+        )
+  return { updated, applied: updated }
+}
